refactor(mobile): share one reorder drag implementation
HealthTrendsSettingsScreen carried a verbatim third copy of the drag
preview hook, the row height and long-press constants, the strides/offsets
memo, and the pan gesture. Three copies of an interaction this fiddly drift
apart as different changes touch each one.

Hoist them next to computeReorderPreviewShift, which the surfaces already
share, and have both settings screens use them:

- useReorderRowPreviewStyle, previously useMealTypeRowDragPreviewStyle
- createReorderRowPanGesture
- useReorderRowGeometry, which also retires the exhaustive-deps disable
  both screens needed around the offsets memo
- resetReorderDragPreview, previously resetMealTypeDragPreview
- REORDER_LONG_PRESS_MS, alongside the exported REORDER_ROW_HEIGHT that
  both screens now use in place of their own ROW_HEIGHT/ROW_GAP pair

Behaviour is unchanged: the gesture is still built per render rather than
memoised, so nothing about how GestureDetector attaches it changes.
